pyproject.toml
[project]
name = "mcp-kql-server"
version = "1.0.0"
description = "MCP server for executing KQL queries with Azure authentication"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-kusto-data==4.0.0",
    "tenacity==8.2.3",
    "werkzeug==2.0.3",
    "mcp",
    "tabulate==0.9.0",
]

[tool.uv]
index-strategy = "unsafe-best-match"

[tool.hatch.build.targets.wheel]
packages = ["mcp_kql_server"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"